+++ kraken/.gitignore
@@ -3,3 +3,4 @@ main.js
 main.d.ts
 coverage
 docs
+lib
