+++ kraken/.npmignore
@@ -5,3 +5,4 @@
 /test
 /.eslintrc.json
 /docs
+/src
